infoTheaterTypePy/requirements.txt
Flask
flask-cors
flask-swagger-ui
psycopg2-binary
python-dotenv
